data_pipeline/frontend/requirements.txt
streamlit
pandas
google-cloud-storage
openpyxl